Passing data
Im adding the ability to pass data between the first loggin and the 
games. Because the games will be a seperate python file. This will also 
have a second effect of not letting someone save data if there has been 
a more recent login with the same account. This uses a second password 
or key that is generated by the program when data is transferred.